refactor: update GeocodingResponse import path and redefine interface
