+++ src/constants/types/weather/weather-client.interface.ts
@@ -66,27 +66,6 @@ export interface Location {
   url: string;
 }
 
-export type GeocodingResponse = {
-  results: {
-    id: number;
-    name: string;
-    latitude: number;
-    longitude: number;
-    elevation: number;
-    feature_code: string;
-    country_code: string;
-    admin1_id: number;
-    admin2_id: number;
-    timezone: string;
-    population: number;
-    country_id: number;
-    country: string;
-    admin1: string;
-    admin2: string;
-  }[];
-  generationtime_ms: number;
-};
-
 export const openMeteoWeatherCodeMap: Record<number, string> = {
   0: 'Clear sky',
   1: 'Mainly clear',
